refactor(core): extract wasm parser wrapper module
